// src/lib/gantt-task-react/types/public-types.ts
export enum ViewMode {
  Hour = "Hour",
  QuarterDay = "Quarter Day",
  HalfDay = "Half Day",
  Day = "Day",
  TwoDays = "Two Days",
  /** ISO-8601 week */
  Week = "Week",
  Month = "Month",
  QuarterYear = "QuarterYear",
  Year = "Year",
}
// ...
export interface Dependency {
  sourceId: string;
  sourceTarget: DateExtremity;
  ownTarget: DateExtremity;
}
// ...
export interface ColorStyles {
  arrowColor: string;
  arrowCriticalColor: string;
  arrowWarningColor: string;
  barLabelColor: string;
  barLabelWhenOutsideColor: string;
  barProgressColor: string;
  barProgressCriticalColor: string;
  barProgressSelectedColor: string;
  barProgressSelectedCriticalColor: string;
  barBackgroundColor: string;
  barBackgroundCriticalColor: string;
  barBackgroundSelectedColor: string;
  barBackgroundSelectedCriticalColor: string;
  groupProgressColor: string;
  groupProgressCriticalColor: string;
  groupProgressSelectedColor: string;
  groupProgressSelectedCriticalColor: string;
  groupBackgroundColor: string;
  groupBackgroundCriticalColor: string;
  groupBackgroundSelectedColor: string;
  groupBackgroundSelectedCriticalColor: string;
  projectProgressColor: string;
  projectProgressCriticalColor: string;
  projectProgressSelectedColor: string;
  projectProgressSelectedCriticalColor: string;
  projectBackgroundColor: string;
  projectBackgroundCriticalColor: string;
  projectBackgroundSelectedColor: string;
  projectBackgroundSelectedCriticalColor: string;
  milestoneBackgroundColor: string;
  milestoneBackgroundCriticalColor: string;
  milestoneBackgroundSelectedColor: string;
  milestoneBackgroundSelectedCriticalColor: string;
  evenTaskBackgroundColor: string;
  oddTaskBackgroundColor: string;
  holidayBackgroundColor: string;
  taskDragColor: string;
  selectedTaskBackgroundColor: string;
  todayColor: string;
  contextMenuBoxShadow: string;
  contextMenuBgColor: string;
  contextMenuTextColor: string;
}
// ...
export type TaskType = "task" | "milestone" | "project";

export interface Task {
  id: string;
  type: TaskType;
  name: string;
  start: Date;
  end: Date;
  /**
   * From 0 to 100
   */
  progress: number;
  assignees?: string[];
  styles?: Partial<ColorStyles>;
  isDisabled?: boolean;
  isRelationDisabled?: boolean;
  /**
   * Project or task
   */
  parent?: string;
  dependencies?: Dependency[];
  hideChildren?: boolean;
  displayOrder?: number;
  comparisonLevel?: number;
}

export interface EmptyTask {
  id: string;
  type: "empty";
  name: string;
  parent?: string;
  comparisonLevel?: number;
  displayOrder?: number;
  isDisabled?: boolean;
  styles?: Partial<ColorStyles>;
}

export type TaskOrEmpty = Task | EmptyTask;